added search input
added search input and reorganized to make a categories filter easier to implement
